requirements.txt
fastapi
uvicorn
httpx
redis
psycopg2-binary
python-dotenv
pydantic
pytest
pytest-asyncio
locust
